Moved Classes to folder: /classes and changed imports
