+++ src/classes/Cart.ts
@@ -1,6 +1,9 @@
 import Product from "./Product"
+
 export default class Cart {
+
   private products: Product[] = []
+
 
   public add(p: Product): void {
     this.products.push(p)
@@ -22,10 +25,9 @@ export default class Cart {
     this.products.splice(index)
   }
 
-  // Marca a todos los productos como no disponibles para futuras compras
-  // Y vacia el carrito.
+  // Marca no disponibles para futuras compras Y vacia el carrito.
   public buyCart(): void {
-    this.products.map(p => p.notAvaliable())
+    this.products.forEach(p => p.notAvaliable())
     this.removeAll()
   }
 }
