+++ src/classes/Members.ts
@@ -1,7 +1,9 @@
 import RegisteredUser from "./RegisteredUser";
 
 export default class Members {
+
   private members: RegisteredUser[] = []
+
 
   public getAll(): RegisteredUser[] {
     return this.members
