+++ src/classes/User.ts
@@ -1,6 +1,8 @@
 export default class User {
+
   readonly _canBuy: boolean = false
   readonly _canComment: boolean = false
+
 
   public get canBuy(): boolean {
     return this._canBuy
